Edited zeroToDevForLoop.sh to fix the loop
The loop didn't work previously but now it does due to the
addition of the new logic. It uses 2 variables to output the final
message. The first variable stores the output of the for loop int
with a string message and the second variables uses the output of
the first along with the word count command to find the first
variables word count and then the final echo statement outputs
the two variables together in the format
"Number <loopCount> : <wordCountOfEntireString>"
example: "Number 0 : 8"