Edited functionWithPositionalArgs.sh
Edited this file to allow for positional arguments when running the
file. The file allows only a specific user as an accepted value but
denies entry if it fails to match. I also changed the name of the
file to have a ".sh" ending since it was initially created without
that ending.