Edited autoBackup.sh & Created backUpDel.sh
Edited autoBackup.sh file to allow the user to input command line
arguments to run the script file. It allows the user to decide
the source and destination directories for the directory to backup.
Created backUpDel.sh to modularize the script file. It contains the
logic to delete backup files after 7 days. A autoBackup_Utils folder
was created to store all the utilities packages that the script
file would use for the project.